microservicios/alert_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
requests
PyJWT
dicttoxml
gunicorn
